WIP almost done with Any.UnmarshalRESP tests
The remaining piece is a multi-step process:

* streaming blob strings must be supported.

* to do this the unmarshalSingle method is going to need to be modified
to not take in a size. This will go hand-in-hand with getting rid of the
byteReaderPool, replacing its usage instead with a wrapper around the
bufio.Reader which stops reading at the first delim and returns EOF.
then unmarshalSingle doesn't need any knowledge of the message size.

* Once that's done another bufio.Reader wrapper can be implemented which
reads off streaming blob strings correctly, and that can be passed into
unmarshalSingle.

This sounds somewhat familiar, I think the original unmarshaling code
was like this but it was removed due to allocation concerns. I should
double check that before continuing.